Volume Control
Adds a volume control mechanism and fader

+++ src/main/java/com/mimlitz/tumbleQ/sound/SoundClip.java
@@ -21,6 +21,6 @@ public interface SoundClip {
 
     List<Integer> getBookmarks();
 
-    void setVolume(int vol);
+    void setVolume(double vol);
 
 }
